CHERRY-PICK PR/819 Implementing filter pipelines and exporting to plugins
It's now possible to plugins to change behaviors/values of the Core UI code.
The main idea is to have pipelines offered by the Core and any
plugins can register to these pipelines and they will be called when the
pipelines run.

The first pipeline implemented here makes it possible to modify any ajax
request made by the UI. Here is how a plugin could register to be called
when this pipeline runs:

```
const {
  PluginHelper,
  PluginMountPoints,
  PipelineNames,
  PipelineStore,
} = window.marathonPluginInterface;

PipelineStore.registerOperator(PipelineNames.PRE_AJAX_REQUEST, (data) => {
  /* Modify "data" and return a new version of it */
});

```